+++ src/main/java/com/harmonia/HarmoniaApplication.java
@@ -1,4 +1,4 @@
-package com.harmoniaApplication.harmonia;
+package com.harmonia;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -11,6 +11,7 @@ public class HarmoniaApplication extends Application {
     @Override
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(HarmoniaApplication.class.getResource("login-view.fxml"));
+        fxmlLoader.setController(new HarmoniaController());
         Scene scene = new Scene(fxmlLoader.load(), 1020, 740);
         stage.setTitle("Harmonia");
         stage.setScene(scene);
@@ -18,6 +19,6 @@ public class HarmoniaApplication extends Application {
     }
 
     public static void main(String[] args) {
-        launch();
+        launch(args);
     }
 }
