+++ src/main/java/com/harmonia/HarmoniaController.java
@@ -1,0 +1,24 @@
+package com.harmonia;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+import javafx.scene.control.TextField;
+import javafx.scene.control.PasswordField;
+
+public class HarmoniaController {
+    @FXML
+    private TextField usernameField;
+
+    @FXML
+    private PasswordField passwordField;
+
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onLoginButtonClick() {
+        String username = usernameField.getText();
+        String password = passwordField.getText();
+        welcomeText.setText("Welcome to Harmonia!");
+    }
+}
