+++ src/main/java/com/harmonia/network/NetworkManager.java
@@ -1,0 +1,5 @@
+package com.harmonia.network;
+
+public class NetworkManager {
+    private NetworkManager networkManager;
+}
