--- src/main/java/com/harmoniaApplication/harmonia/HarmoniaController.java
@@ -1,14 +1,0 @@
-package com.harmoniaApplication.harmonia;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HarmoniaController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onLoginButtonClick() {
-        welcomeText.setText("Welcome to Harmonia!");
-    }
-}
