--- src/main/java/com/harmoniaApplication/harmonia/azure/NetworkManager.java
@@ -1,4 +1,0 @@
-package com.harmoniaApplication.harmonia.azure;
-
-public class NetworkManager {
-}
