+++ src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module com.example.harmonia {
+module com.harmonia {
     requires javafx.controls;
     requires javafx.fxml;
     requires javafx.web;
@@ -10,6 +10,6 @@ module com.example.harmonia {
     requires org.kordamp.bootstrapfx.core;
     requires eu.hansolo.tilesfx;
 
-    opens com.harmoniaApplication.harmonia to javafx.fxml;
-    exports com.harmoniaApplication.harmonia;
+    opens com.harmonia to javafx.fxml;
+    exports com.harmonia;
 }
