+++ src/main/resources/com/harmonia/login-view.fxml
@@ -3,14 +3,20 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.VBox?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.PasswordField?>
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.harmoniaApplication.harmonia.HarmoniaController">
+      fx:controller="com.harmonia.HarmoniaController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="30.0"/>
     </padding>
 
     <Label fx:id="welcomeText"/>
+    <Label text="Username:" />
+    <TextField fx:id="usernameField" />
+    <Label text="Password:" />
+    <PasswordField fx:id="passwordField" />
     <Button text="Login" onAction="#onLoginButtonClick"/>
 </VBox>
